move webworker code to own module
